+++ python/cugraph/opg/link_analysis/mg_pagerank_wrapper.pyx
@@ -39,6 +39,7 @@ def mg_pagerank(input_df, local_data, handle, alpha=0.85, max_iter=100, tol=1.0e
     local_offset = dst.min() # Find using local data offset array
     dst = dst - local_offset
     num_local_verts = dst.max() + 1 # Find using local data verts array
+    num_local_edges = len(src)
     cdef uintptr_t c_local_verts = local_data['verts'].__array_interface__['data'][0]
     cdef uintptr_t c_local_edges = local_data['edges'].__array_interface__['data'][0]
     cdef uintptr_t c_local_offsets = local_data['offsets'].__array_interface__['data'][0]
@@ -66,13 +67,15 @@ def mg_pagerank(input_df, local_data, handle, alpha=0.85, max_iter=100, tol=1.0e
     cdef GraphCSCView[int,int,double] graph_double
 
     if (df['pagerank'].dtype == np.float32):
-        graph_float = GraphCSCView[int,int,float](<int*>c_offsets, <int*>c_indices, <float*>c_weights, num_verts, num_edges)
+        graph_float = GraphCSCView[int,int,float](<int*>c_offsets, <int*>c_indices, <float*>c_weights, num_verts, num_local_edges)
         graph_float.set_local_data(<int*>c_local_verts, <int*>c_local_edges, <int*>c_local_offsets)
+        graph_float.set_handle(handle_)
         c_pagerank.pagerank[int,int,float](handle_[0], graph_float, <float*> c_pagerank_val, 0, <int*> NULL, <float*> NULL,
                                <float> alpha, <float> tol, <int> max_iter, <bool> 0)
     else:
-        graph_double = GraphCSCView[int,int,double](<int*>c_offsets, <int*>c_indices, <double*>c_weights, num_verts, num_edges)
+        graph_double = GraphCSCView[int,int,double](<int*>c_offsets, <int*>c_indices, <double*>c_weights, num_verts, num_local_edges)
         graph_double.set_local_data(<int*>c_local_verts, <int*>c_local_edges, <int*>c_local_offsets)
+        graph_double.set_handle(handle_)
         c_pagerank.pagerank[int,int,double](handle_[0], graph_double, <double*> c_pagerank_val, 0, <int*> NULL, <double*> NULL,
                             <float> alpha, <float> tol, <int> max_iter, <bool> 0)